requirements.txt
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
celery==3.1.18
celery-with-redis==3.0
cffi==1.2.1
cryptography==1.0.1
Django==1.8.4
django-celery==3.1.16
django-filter==0.11.0
djangorestframework==3.2.4
enum34==1.0.4
geopy==1.11.0
idna==2.0
ipaddress==1.0.14
kombu==3.0.26
lxml==3.4.4
Markdown==2.6.2
ndg-httpsclient==0.4.0
py==1.4.30
pyasn1==0.1.8
pycparser==2.14
pyOpenSSL==0.15.1
pytest==2.8.1
pytest-django==2.8.0
pytz==2015.6
redis==2.10.3
requests==2.7.0
requests-mock==0.6.0
six==1.9.0
wheel==0.24.0
